pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "werkzeug>=3.1.6",
]
